fix pagination styling and move outside to explore page
